bigdecimale, refaktoryzacja, wip mocka

+++ src/main/java/pl/timsus/recycoinbackend/distributor/dto/IdentifiersValueDto.java
@@ -2,18 +2,20 @@ package pl.timsus.recycoinbackend.distributor.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class IdentifiersValueDto extends IdentifiersDto{
-    private final Double value;
+    private final BigDecimal value;
 
     public IdentifiersValueDto(
             @JsonProperty(required = true) Integer distributorId,
             @JsonProperty(required = true) Integer clientId,
-            @JsonProperty(required = true) Double value) {
+            @JsonProperty(required = true) BigDecimal value) {
         super(distributorId, clientId);
         this.value = value;
     }
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 }
